Clean up code: All #includes are in main.cpp ++ seperate enum definitions from settings file (definitions.h)
Former-commit-id: 730f28debf83844fd79fc4b44d88b91b94421ea2

+++ Code/BA/Headers/fast_updates.h
@@ -1,11 +1,9 @@
-#include "init.h"
-#include "full_statistics.h"
 
 // BA edits: argument 6: changed pass-by-reference to pass-by-value
 // BA edtis (24/5/16): used templates
 
 template<class T>
-void fast_updates(T **BASIS, T *targets, T *coefficients, int N, int M, double noiseStd, T *errors, T **PSI,bool calc, int change_in_ML_threshold, bool printToCOut)
+void fast_updates(T **BASIS, T *targets, T *coefficients, int N, int M, double noiseStd, T *errors, T **PSI,bool calc, double change_in_ML_threshold, bool printToCOut)
 {
     if (printToCOut) {
         std::cout << "RVM: \t" << std::flush;
